+++ src/collective/cover/upgrades.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from collective.cover.config import PROJECTNAME
-from Products.CMFCore.utils import getToolByName
+from plone import api
 
 import logging
 
@@ -14,7 +14,7 @@ def issue_201(context, logger=None):
         logger = logging.getLogger(PROJECTNAME)
 
     # first we take care of the CSS registry
-    css_tool = getToolByName(context, 'portal_css')
+    css_tool = api.portal.get_tool(name='portal_css')
     old_id = '++resource++collective.cover/bootstrap.min.css'
     if old_id in css_tool.getResourceIds():
         css_tool.unregisterResource(old_id)
@@ -25,7 +25,7 @@ def issue_201(context, logger=None):
         logger.debug('"{0}" resource not found in portal_css'.format(old_id))
 
     # now we mess with the JS registry
-    js_tool = getToolByName(context, 'portal_javascripts')
+    js_tool = api.portal.get_tool(name='portal_javascripts')
     old_id = '++resource++collective.cover/bootstrap.min.js'
     new_id = '++resource++collective.js.bootstrap/js/bootstrap.min.js'
     if old_id in js_tool.getResourceIds():
@@ -54,7 +54,7 @@ def issue_303(context, logger=None):
                   'tiny_mce.js',
                   'tiny_mce_init.js']
 
-    js_tool = getToolByName(context, 'portal_javascripts')
+    js_tool = api.portal.get_tool(name='portal_javascripts')
     for id in js_tool.getResourceIds():
         if id in FIX_JS_IDS:
             js = js_tool.getResource(id)